Add rmagick gem
It's an alpine package, whichis super-nice since then I don't have to
compile it myself. Gems suck in that regard. There must be a way to
"compile, package, distribute" because having an entire development
suite on production systems is just stupidly bloated

Also copy in the right split file (ruby, not shell, thx)